clients/maintainer: move more into layout
This makes the navigation between the pages smoother
