+++ clients/apps/web/src/app/maintainer/[organization]/finance/layout.tsx
@@ -1,9 +1,9 @@
-import DashboardLayout from '@/components/Layout/DashboardLayout'
+import { DashboardBody } from '@/components/Layout/DashboardLayout'
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <>
-      <DashboardLayout showSidebar={true}>{children}</DashboardLayout>
+      <DashboardBody>{children}</DashboardBody>
     </>
   )
 }
