+++ clients/apps/web/src/app/maintainer/layout.tsx
@@ -1,9 +1,12 @@
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
+import DashboardLayout from '@/components/Layout/DashboardLayout'
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <Gatekeeper>
-      <>{children}</>
+      <DashboardLayout>
+        <>{children}</>
+      </DashboardLayout>
     </Gatekeeper>
   )
 }
